requirements.txt
Django==2.2.1
django-allauth==0.39.1
djangorestframework==3.9.4
django-model-utils==3.1.2
django-mptt==0.10.0
django-crispy-forms==1.7.2
django-ipware==2.1.0
geoip2==2.9.0
django-taggit==1.1.0
django-debug-toolbar==1.11
bleach==3.1.0
Markdown==3.1
Pygments==2.4.0
easy-thumbnails==2.6
django-import-export==1.2.0
celery==4.3.0
python-memcached==1.59
django-otp==0.6.0
qrcode==6.1
django-disqus==0.5
WeasyPrint==47
Pillow==6.0.0
